Use systemd watchdog (WIP)
This is a work-in-progress to add systemd support. When valid data is
received from a device, it notifies the watchdog, allowing the process
to be terminated if for example bluetooth connection goes bad or
service is stalled for any other reason. To activate, add the
following to [Service] part of a unit:

WatchdogSec=1min

Which restarts if nothing is received in a minute.